Feat: Footer Tag 검색 컴포넌트 분리 및 디자인 변경
