+++ src/components/UI/Slide.js
@@ -1,0 +1,67 @@
+import { useEffect } from "react";
+import styled from "styled-components";
+
+function Slide({ children, setSlideIsOpen }) {
+  useEffect(() => {
+    document.body.style = `overflow: hidden`;
+    return () => (document.body.style = `overflow: auto`);
+  }, []);
+
+  return (
+    <SlideBack onClick={() => setSlideIsOpen(false)}>
+      <SlideBox onClick={(e) => e.stopPropagation()}>{children}</SlideBox>
+    </SlideBack>
+  );
+}
+
+export default Slide;
+
+const SlideBack = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+`;
+
+const SlideBox = styled.div`
+  background-color: var(--color-white);
+  border-radius: 1.5rem 1.5rem 0 0;
+  box-shadow: 0 0 10px rgb(0 0 0 / 30%);
+  text-align: center;
+  width: 100%;
+  height: 85%;
+  padding: 1rem 2rem 2rem 2rem;
+  transition: all 600ms cubic-bezier(0.86, 0, 0.07, 1);
+  z-index: 2;
+  overflow: scroll;
+
+  ::-webkit-scrollbar {
+    width: 0.3rem;
+  }
+  ::-webkit-scrollbar-thumb {
+    background-color: var(--color-orange);
+    border-radius: 10px;
+  }
+  ::-webkit-scrollbar-track {
+    margin-top: 1.3rem;
+    background-color: var(--color-white);
+  }
+
+  animation-name: slideMotion;
+  animation-duration: 1s;
+
+  @keyframes slideMotion {
+    from {
+      margin-top: 100vh;
+    }
+    to {
+      margin-top: 0%;
+    }
+  }
+`;
