+++ src/components/common/List.jsx
@@ -5,15 +5,7 @@ import Recipe from "../posts/Recipe";
 const List = ({ list, likeToggle, listName, ...props }) => {
   return (
     <ListContainer {...props}>
-      {list &&
-        list?.map((item) => (
-          <Recipe
-            post={item}
-            key={item.postId}
-            likeToggle={likeToggle}
-            {...props}
-          />
-        ))}
+      {list && list?.map((item) => <Recipe post={item} key={item.postId} likeToggle={likeToggle} {...props} />)}
     </ListContainer>
   );
 };
